% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_in_categorical_variable.R
\name{my_in_categorical_variable}
\alias{my_in_categorical_variable}
\title{my_in_categorical_variable}
\usage{
my_in_categorical_variable(
  name_in,
  df_in = IN,
  levels_in = c(),
  levels_in_to_exclude = c("NA", NA),
  name_out = "",
  df_out = DF,
  label_out = "",
  levels_out = c()
)
}
\arguments{
\item{name_in}{Name of the variable in the input data-frame. String. Default: None.}

\item{df_in}{Input data-frame (with the 'name_in' variable). Data-frame. Default: IN.}

\item{levels_in}{Levels to be used in the in the original variable (should be in the desired order; if absent, available levels are used). String vector. Default: c().}

\item{levels_in_to_exclude}{Levels to be excluded. String vector. Default: c('NA', NA).}

\item{name_out}{Name for the variable in the resulting data-frame (if absent, 'name_in' is used). String. Default: ''.}

\item{df_out}{Output data-frame (possibly without the 'name_out' variable). Data-frame. Default: DF.}

\item{label_out}{Label (using 'Hmisc::label()') for the variable in the resulting data-frame (if absent, 'Hmisc:label()' of 'name_in' is used, possibly set to 'name_in'). String. Default: ''.}

\item{levels_out}{Levels to be used in the in the resulting data-frame (should correspond to those indicated in 'levels_in'; if absent, original levels are used). String vector. Default: c().}
}
\value{
A data-frame with the main variable and associated dummy variables (column-binded to 'df_out').
}
\description{
Function to insert a categorical variable from a data-frame into another data-frame, creating dummy variables (one for each level).
}
