% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean_ci.R
\name{my_mean_ci}
\alias{my_mean_ci}
\title{my_mean_ci}
\usage{
my_mean_ci(x, confidence = 0.95, distribution = "z")
}
\arguments{
\item{x}{Numeric vector. Default: None.}

\item{confidence}{Confidence interval desired. Numeric value. Default: 0.950.}

\item{distribution}{z-distributionn or t-distribution. For small or poorly distributed samples, use t-distribution. String. Default: 'z'.}
}
\value{
A vector: c(y, ymin, ymax).
}
\description{
Function to obtain the confidence interval of the mean. Using: 'psych' package.
}
