% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice.R
\name{my_nice}
\alias{my_nice}
\title{my_nice}
\usage{
my_nice(
  value,
  decimals = 3,
  text = "",
  with_equal_sign = FALSE,
  with_sign = TRUE,
  min_value = -Inf,
  max_value = Inf,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 3.}

\item{text}{Text to name the value. String. Default: ''.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is some text in 'text'). Boolean. Default: FALSE.}

\item{with_sign}{Mark of +/- in front of the number. Boolean. Default: TRUE.}

\item{min_value}{Minimum permissible/possible value. If 'value' is lower than it, will be rounded to this. Numeric value. Default: -Inf.}

\item{max_value}{Maximum permissible/possible value. If 'value' is higher than it, will be rounded to this. Numeric value. Default: Inf.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the number with required decimals, possibly with the name of the value and/or with several ornaments.
}
\description{
Function to represent a number rounded to the required decimal places.
From a numeric value to a string.
}
