% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_p.R
\name{my_nice_p}
\alias{my_nice_p}
\title{my_nice_p}
\usage{
my_nice_p(
  value,
  decimals = 3,
  with_p = TRUE,
  with_equal_sign = FALSE,
  with_stars = TRUE,
  multiple_stars = TRUE,
  alpha = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001),
  give_only_stars = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 3.}

\item{with_p}{Shows 'p' (and equal, or unequal, sign) before the p-value. Boolean. Default: TRUE.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is 'p'). Boolean. Default: FALSE.}

\item{with_stars}{Shows an asterisk if statistically significant. Boolean. Default: TRUE.}

\item{multiple_stars}{Shows one-to-three asterisks if statistically significant (depending on level, see 'multiple_alphas'). Boolean. Default: TRUE.}

\item{alpha}{Statistical significance. Numeric value. Default: 0.050.}

\item{multiple_alphas}{Numeric vector with three levels of statistical significance (for multiple asterisks). Numeric vector. Default: c(0.050, 0.010, 0.001).}

\item{give_only_stars}{Returns only the asterisks (without the p-value). Boolean. Default = FALSE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the p-value with required decimals, possibly with several ornaments.
}
\description{
Function to represent p-values.
Includes asterisks for statistical significance.
}
