% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_percent.R
\name{my_nice_percent}
\alias{my_nice_percent}
\title{my_nice_percent}
\usage{
my_nice_percent(
  value,
  decimals = 1,
  text = "",
  percent_sign = TRUE,
  per_level = "100",
  min_value = -500,
  max_value = 500,
  with_equal_sign = FALSE,
  with_sign = TRUE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 1.}

\item{text}{Text to name the value. String. Default: ''.}

\item{percent_sign}{Uses the percentage sign. Boolean. Default: TRUE.}

\item{per_level}{Uses the percentage sign for 1/100, 1/1000, or 1/10000. String. Default: '100'.}

\item{min_value}{Minimum permissible/possible value. If 'value' is lower than it, will be rounded to this. Numeric value. Default: -500.}

\item{max_value}{Maximum permissible/possible value. If 'value' is higher than it, will be rounded to this. Numeric value. Default: 500.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is some text in 'text'). Boolean. Default: FALSE.}

\item{with_sign}{Mark of +/- in front of the number. Boolean. Default: TRUE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the percentage with required decimals, possibly with several ornaments.
}
\description{
Function to represent percentages.
}
