% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_r.R
\name{my_nice_r}
\alias{my_nice_r}
\title{my_nice_r}
\usage{
my_nice_r(
  value,
  decimals = 3,
  with_r = TRUE,
  spearman = FALSE,
  with_equal_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 3.}

\item{with_r}{Shows 'r'/'rS' (and equal, or unequal, sign) before the value. Boolean. Default: TRUE.}

\item{spearman}{Shows 'rS' instead of 'r' if there is 'with_r'. Boolean. Default: FALSE.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is 'with_r'). Boolean. Default: FALSE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the correlation with required decimals, possibly with several ornaments.
}
\description{
Function to represent correlations.
}
