% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_ss.R
\name{my_nice_ss}
\alias{my_nice_ss}
\title{my_nice_ss}
\usage{
my_nice_ss(
  value,
  decimals = 1,
  text = "SS",
  min_value = 25,
  max_value = 175,
  with_equal_sign = FALSE,
  with_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 1.}

\item{text}{Text to name the value. String. Default: 'SS'.}

\item{min_value}{Minimum permissible/possible value. If 'value' is lower than it, will be rounded to this. Numeric value. Default: 25.}

\item{max_value}{Maximum permissible/possible value. If 'value' is higher than it, will be rounded to this. Numeric value. Default: 175.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is some text in 'text'). Boolean. Default: FALSE.}

\item{with_sign}{Mark of +/- in front of the number. Boolean. Default: FALSE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the SS-value with required decimals, possibly with several ornaments.
}
\description{
Function to represent SS-values (mean: 100, SD: 15).
}
