% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_t.R
\name{my_nice_t}
\alias{my_nice_t}
\title{my_nice_t}
\usage{
my_nice_t(
  value,
  decimals = 2,
  text = "T",
  min_value = 0,
  max_value = 100,
  with_equal_sign = FALSE,
  with_sign = FALSE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 2.}

\item{text}{Text to name the value. String. Default: 'T'.}

\item{min_value}{Minimum permissible/possible value. If 'value' is lower than it, will be rounded to this. Numeric value. Default: 0.}

\item{max_value}{Maximum permissible/possible value. If 'value' is higher than it, will be rounded to this. Numeric value. Default: 100.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is some text in 'text'). Boolean. Default: FALSE.}

\item{with_sign}{Mark of +/- in front of the number. Boolean. Default: FALSE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the T-value with required decimals, possibly with several ornaments.
}
\description{
Function to represent T-values (mean: 50, SD: 10).
}
