% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_nice_z.R
\name{my_nice_z}
\alias{my_nice_z}
\title{my_nice_z}
\usage{
my_nice_z(
  value,
  decimals = 3,
  text = "",
  standard_mean = 0,
  standard_sd = 1,
  min_z = -5,
  max_z = 5,
  with_equal_sign = FALSE,
  with_sign = TRUE,
  void_string = "-"
)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 3.}

\item{text}{Text to name the value. String. Default: ''.}

\item{standard_mean}{Mean of the standard score. Numeric value. Default: 0.}

\item{standard_sd}{SD of the standard score. Numeric value. Default: 1.}

\item{min_z}{Minimum (in SD from the mean) permissible/possible value. If 'value' is lower than it, will be rounded to this. Numeric value. Default: -5.}

\item{max_z}{Maximum (in SD from the mean) permissible/possible value. If 'value' is higher than it, will be rounded to this. Numeric value. Default: 5.}

\item{with_equal_sign}{Equal, or unequal, sign (always present if there is some text in 'text'). Boolean. Default: FALSE.}

\item{with_sign}{Mark of +/- in front of the number. Boolean. Default: TRUE.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}
}
\value{
A string for the standard value with required decimals, possibly with several ornaments.
}
\description{
Function for representing z-values (but also applicable to other standardisations).
E.g., To represent T-values, use: 'standard_mean = 50', 'standard_sd = 10'.
}
