% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_paired_ttest.R
\name{my_paired_ttest}
\alias{my_paired_ttest}
\title{my_paired_ttest}
\usage{
my_paired_ttest(
  y,
  time,
  void_string = "-",
  alpha_value = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001),
  wise = TRUE,
  direction = "Stable"
)
}
\arguments{
\item{y}{Numeric vector. Default: None.}

\item{time}{Factor vector (with 2 levels). Default: None.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}

\item{alpha_value}{Statistical significance. Numeric value. Default: 0.050.}

\item{multiple_alphas}{Numeric vector with three levels of statistical significance (for multiple asterisks). Numeric vector. Default: c(0.050, 0.010, 0.001).}

\item{wise}{Boolean, if true, the most appropriate test is used according to the data. Default: TRUE}

\item{direction}{Specifying the alternative hypothesis (using: 'Stable', 'Increase', 'Decrease'). String. Default: 'Stable'.}
}
\value{
A list with results: 'test' (string, with results of the within-subjects t-test), 'p_value' (numeric, the value of p associated with the test), 'significance' (string, with an asterisk for statistically significant results), 'comparison' (string, comparisons between levels of the time variable marked when the result is statistically significant), 'es' (string, effect-size for statistically significant results), 'times' (string, mean and SD for the levels of the time variable), 'times_pairs' (list of vectors, every vector reports two times corresponding to post-hoc comparisons), 'times_pairs_p' (vector of numeric, significances corresponding to post-hoc comparisons).
}
\description{
Function to do a within-subjects Test t.
}
