% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_round.R
\name{my_round}
\alias{my_round}
\title{my_round}
\usage{
my_round(value, decimals = 3)
}
\arguments{
\item{value}{Numeric value. Default: None.}

\item{decimals}{Number of decimals. Numeric value. Default: 3.}
}
\value{
A string for the number with required decimals.
}
\description{
Function to represent a number rounded to the required decimal places.
From a numeric value to a string.
}
