% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{my_save_table}
\alias{my_save_table}
\title{my_save_table}
\usage{
my_save_table(
  what,
  file_name = "DF",
  where = getwd(),
  row_names = FALSE,
  with_rdata = TRUE,
  with_csv = FALSE,
  fileEncoding = "latin1"
)
}
\arguments{
\item{what}{Data-frame to be saved. Data-frame. Default: None.}

\item{file_name}{Name by which the file will be saved. String. Default: 'DF'.}

\item{where}{Destination folder. String. Default: current folder (with 'getwd()').}

\item{row_names}{Use of row names. Boolean. Default: FALSE.}

\item{with_rdata}{Save as 'RData'. Boolean. Default: TRUE.}

\item{with_csv}{Save as 'csv' (both standard and Italian-version 'csv' are saved). Boolean. Default: FALSE.}

\item{fileEncoding}{File encoding. String. Default: 'latin1'.}
}
\value{
Nothing.
}
\description{
Function to save a data-frame in 'xlsx' and/or 'RData' and/or 'csv'. Using: 'xlsx' package.
}
