% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_variable_classification.R
\name{my_variable_classification}
\alias{my_variable_classification}
\title{my_variable_classification}
\usage{
my_variable_classification(variable = "", df = DF)
}
\arguments{
\item{A}{a data.frame (where the variable to classify is). Data-frame. Default: DF.}
}
\value{
A string: 'Unknown', 'Continuous', 'Poor continuous', 'Dummy', 'Empty continuous', 'Dichotomous', 'Polytomous', 'High polytomous', 'Empty categorical', 'Date', 'Empty'
}
\description{
Function to classify a variable (for further analyses).
}
