% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_in_categorical_variable.R
\name{my_in_categorical_variable}
\alias{my_in_categorical_variable}
\title{my_in_categorical_variable}
\usage{
my_in_categorical_variable(
  name_in,
  df_in = IN,
  levels_in_to_exclude = c("NA"),
  name_out = "",
  levels_out = c()
)
}
\arguments{
\item{name_in}{Name of the variable in the input data.frame. String. Default: None.}

\item{df_in}{Input data.frame (with the 'name_in' variable). Data.frame. Default: IN.}

\item{levels_in_to_exclude}{Levels to be excluded. String vector. Default: c('NA').}

\item{name_out}{Name for the variable in the resulting data.frame (if absent, 'name_in' is used). String. Default: ''.}

\item{levels_out}{Levels to be used in the in the resulting data.frame (should be in the same number and order as the original levels; if absent, original levels are used). String vector. Default: c().}
}
\value{
A data.frame with the main variable and associated dummy variables.
}
\description{
Function to insert a categorical variable, creating dummy variables (one for each level).
}
