% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_spearman_r.R
\name{my_spearman_r}
\alias{my_spearman_r}
\title{my_spearman_r}
\usage{
my_spearman_r(
  y,
  x,
  void_string = "-",
  alpha_value = 0.05,
  multiple_alphas = c(0.05, 0.01, 0.001)
)
}
\arguments{
\item{y}{Numeric vector. Default: None.}

\item{x}{Numeric vector. Default: None.}

\item{void_string}{String to be used if the number cannot be represented correctly. String. Default: '-'.}

\item{alpha_value}{Statistical significance. Numeric value. Default: 0.050.}

\item{multiple_alphas}{Numeric vector with three levels of statistical significance (for multiple asterisks). Numeric vector. Default: c(0.050, 0.010, 0.001).}
}
\value{
A list with results: 'test' (string, with results of the Pearson's correlation), 'p_value' (numeric, the value of p associated with the test), 'significance' (string, with an asterisk for statistically significant results), 'comparison' (string, reporting the direction of the correlation), 'es' (string, effect-size for statistically significant results), 'groups' (string, actually a void string).
}
\description{
Function to do a Pearson's correlation.
}
